requirements.txt
Django~=3.2.4
django-crispy-forms~=1.12.0